move tunable values to a separate object delete unused constants
