+++ src/main/kotlin/com/team4099/robot2025/subsystems/elevator/Elevator.kt
@@ -1,24 +1,14 @@
 package com.team4099.robot2025.subsystems.elevator
 
 import com.team4099.lib.hal.Clock
-import com.team4099.lib.logging.LoggedTunableNumber
-import com.team4099.lib.logging.LoggedTunableValue
 import com.team4099.robot2025.config.constants.ElevatorConstants
 import com.team4099.robot2025.util.CustomLogger
 import edu.wpi.first.wpilibj.RobotBase
-import org.team4099.lib.controller.TrapezoidProfile
-import org.team4099.lib.units.base.Meter
 import org.team4099.lib.units.base.inInches
 import org.team4099.lib.units.base.inches
-import org.team4099.lib.units.base.meters
 import org.team4099.lib.units.base.seconds
 import org.team4099.lib.units.derived.ElectricalPotential
 import org.team4099.lib.units.derived.inVolts
-import org.team4099.lib.units.derived.inVoltsPerInch
-import org.team4099.lib.units.derived.inVoltsPerInchPerSecond
-import org.team4099.lib.units.derived.inVoltsPerInchSeconds
-import org.team4099.lib.units.derived.perInch
-import org.team4099.lib.units.derived.perInchSeconds
 import org.team4099.lib.units.derived.volts
 import org.team4099.lib.units.inInchesPerSecond
 import org.team4099.lib.units.perSecond
@@ -26,123 +16,6 @@ import com.team4099.robot2025.subsystems.superstructure.Request.ElevatorRequest 
 
 class Elevator(val io: ElevatorIO) {
   val inputs = ElevatorIO.ElevatorInputs()
-
-  private val kP =
-    LoggedTunableValue("Elevator/kP", Pair({ it.inVoltsPerInch }, { it.volts.perInch }))
-  private val kI =
-    LoggedTunableValue(
-      "Elevator/kI", Pair({ it.inVoltsPerInchSeconds }, { it.volts.perInchSeconds })
-    )
-  private val kD =
-    LoggedTunableValue(
-      "Elevator/kD", Pair({ it.inVoltsPerInchPerSecond }, { it.volts / 1.0.inches.perSecond })
-    )
-
-  private val slot1kP =
-    LoggedTunableValue("Elevator/slot1kP", Pair({ it.inVoltsPerInch }, { it.volts.perInch }))
-  private val slot1kI =
-    LoggedTunableValue(
-      "Elevator/slot1kI", Pair({ it.inVoltsPerInchSeconds }, { it.volts.perInchSeconds })
-    )
-  private val slot1kD =
-    LoggedTunableValue(
-      "Elevator/slot1kD",
-      Pair({ it.inVoltsPerInchPerSecond }, { it.volts / 1.0.inches.perSecond })
-    )
-
-  object TunableElevatorHeights {
-    val enableElevator =
-      LoggedTunableNumber(
-        "Elevator/enableMovementElevator", if (ElevatorConstants.ENABLE_ELEVATOR) 1.0 else 0.0
-      )
-
-    val minPosition =
-      LoggedTunableValue(
-        "Elevator/minPosition",
-        ElevatorConstants.DOWNWARDS_EXTENSION_LIMIT,
-        Pair({ it.inInches }, { it.inches })
-      )
-
-    val maxPosition =
-      LoggedTunableValue(
-        "Elevator/maxPosition",
-        ElevatorConstants.UPWARDS_EXTENSION_LIMIT,
-        Pair({ it.inInches }, { it.inches })
-      )
-
-    val testPosition =
-      LoggedTunableValue(
-        "Elevator/testPosition",
-        ElevatorConstants.L2_HEIGHT,
-        Pair({ it.inInches }, { it.inches })
-      )
-
-    val openLoopExtendVoltage =
-      LoggedTunableValue(
-        "Elevator/openLoopExtendVoltage",
-        ElevatorConstants.OPEN_LOOP_EXTEND_VOLTAGE,
-        Pair({ it.inVolts }, { it.volts })
-      )
-    val openLoopRetractVoltage =
-      LoggedTunableValue(
-        "Elevator/openLoopRetractVoltage",
-        ElevatorConstants.OPEN_LOOP_RETRACT_VOLTAGE,
-        Pair({ it.inVolts }, { it.volts })
-      )
-
-    val slamVelocity =
-      LoggedTunableValue(
-        "Elevator/slamVelocity",
-        ElevatorConstants.SLAM_VELOCITY,
-        Pair({ it.inInchesPerSecond }, { it.inches.perSecond })
-      )
-
-    val L1Height =
-      LoggedTunableValue(
-        "Elevator/L1Height", ElevatorConstants.L1_HEIGHT, Pair({ it.inInches }, { it.inches })
-      )
-
-    val L2Height =
-      LoggedTunableValue(
-        "Elevator/L2Height", ElevatorConstants.L2_HEIGHT, Pair({ it.inInches }, { it.inches })
-      )
-
-    val L3Height =
-      LoggedTunableValue(
-        "Elevator/L3Height", ElevatorConstants.L3_HEIGHT, Pair({ it.inInches }, { it.inches })
-      )
-
-    val L4Height =
-      LoggedTunableValue(
-        "Elevator/L4Height", ElevatorConstants.L4_HEIGHT, Pair({ it.inInches }, { it.inches })
-      )
-
-    val algaeLowHeight =
-      LoggedTunableValue(
-        "Elevator/algaeLowHeight",
-        ElevatorConstants.ALGAE_LOW_HEIGHT,
-        Pair({ it.inInches }, { it.inches })
-      )
-
-    val algaeHighHeight =
-      LoggedTunableValue(
-        "Elevator/algaeHighHeight",
-        ElevatorConstants.ALGAE_LOW_HEIGHT,
-        Pair({ it.inInches }, { it.inches })
-      )
-
-    val yPos =
-      LoggedTunableValue(
-        "Elevator/yPos",
-        0.0.meters,
-      )
-
-    val y1Pos =
-      LoggedTunableValue(
-        "Elevator/y1Pos",
-        0.0.meters,
-      )
-  }
 
   var targetVoltage: ElectricalPotential = 0.0.volts
 
@@ -180,13 +53,7 @@ class Elevator(val io: ElevatorIO) {
   private var lastRequestedVelocity = -9999.inches.perSecond
   private var lastRequestedVoltage = (-9999).volts
 
-  private var timeProfileGeneratedAt = Clock.fpgaTime
   private var lastHomingStatorCurrentTripTime = Clock.fpgaTime
-
-  private var elevatorConstraints: TrapezoidProfile.Constraints<Meter> =
-    TrapezoidProfile.Constraints(
-      ElevatorConstants.MAX_VELOCITY, ElevatorConstants.MAX_ACCELERATION
-    )
 
   private val isAtTargetedPosition: Boolean
     get() =
@@ -195,7 +62,7 @@ class Elevator(val io: ElevatorIO) {
           (inputs.elevatorPosition - elevatorPositionTarget).absoluteValue <=
           ElevatorConstants.ELEVATOR_TOLERANCE
         ) ||
-        (TunableElevatorHeights.enableElevator.get() != 1.0)
+        (ElevatorTunableValues.TunableElevatorHeights.enableElevator.get() != 1.0)
 
   val canContinueSafely: Boolean
     get() =
@@ -211,34 +78,34 @@ class Elevator(val io: ElevatorIO) {
         )
 
   init {
-    TunableElevatorHeights
+    ElevatorTunableValues.TunableElevatorHeights
 
     if (RobotBase.isReal()) {
       isHomed = false
 
-      kP.initDefault(ElevatorConstants.PID.REAL_KP)
-      kI.initDefault(ElevatorConstants.PID.REAL_KI)
-      kD.initDefault(ElevatorConstants.PID.REAL_KD)
+      ElevatorTunableValues.slot0kP.initDefault(ElevatorConstants.PID.REAL_KP)
+      ElevatorTunableValues.slot0kI.initDefault(ElevatorConstants.PID.REAL_KI)
+      ElevatorTunableValues.slot0kD.initDefault(ElevatorConstants.PID.REAL_KD)
     } else {
       isHomed = true
 
-      kP.initDefault(ElevatorConstants.PID.SIM_KP)
-      kI.initDefault(ElevatorConstants.PID.SIM_KI)
-      kD.initDefault(ElevatorConstants.PID.SIM_KD)
+      ElevatorTunableValues.slot0kP.initDefault(ElevatorConstants.PID.SIM_KP)
+      ElevatorTunableValues.slot0kI.initDefault(ElevatorConstants.PID.SIM_KI)
+      ElevatorTunableValues.slot0kD.initDefault(ElevatorConstants.PID.SIM_KD)
 
-      io.configFirstStagePID(kP.get(), kI.get(), kD.get())
-      io.configSecondStagePID(slot1kP.get(), slot1kI.get(), slot1kD.get())
+      io.configFirstStagePID(ElevatorTunableValues.slot0kP.get(), ElevatorTunableValues.slot0kI.get(),ElevatorTunableValues.slot0kD.get())
+      io.configSecondStagePID(ElevatorTunableValues.slot1kP.get(), ElevatorTunableValues.slot1kI.get(), ElevatorTunableValues.slot1kD.get())
     }
   }
 
   fun periodic() {
     io.updateInputs(inputs)
 
-    if (kP.hasChanged() || kI.hasChanged() || kD.hasChanged()) {
-      io.configFirstStagePID(kP.get(), kI.get(), kD.get())
+    if (ElevatorTunableValues.slot0kP.hasChanged() || ElevatorTunableValues.slot0kI.hasChanged() || ElevatorTunableValues.slot0kD.hasChanged()) {
+      io.configFirstStagePID(ElevatorTunableValues.slot0kP.get(), ElevatorTunableValues.slot0kI.get(), ElevatorTunableValues.slot0kD.get())
     }
-    if (slot1kD.hasChanged() || slot1kP.hasChanged() || slot1kI.hasChanged()) {
-      io.configSecondStagePID(slot1kP.get(), slot1kI.get(), slot1kD.get())
+    if (ElevatorTunableValues.slot1kD.hasChanged() || ElevatorTunableValues.slot1kP.hasChanged() || ElevatorTunableValues.slot1kI.hasChanged()) {
+      io.configSecondStagePID(ElevatorTunableValues.slot1kP.get(), ElevatorTunableValues.slot1kI.get(), ElevatorTunableValues.slot1kD.get())
     }
 
     CustomLogger.processInputs("Elevator", inputs)
